fix: harden reimport data integrity boundaries (#2621)
## Summary

Land the fresh-rebuild integrity campaign slice covering bounded replay
handoffs, CAS/material durability, manifest-backed replay,
source-material lifecycle, and snapshot/NATS safeguards.

## Problem

The pre-wipe audit found several ways a large re-import could retain
unbounded in-memory scope state, lose replay authority after source
removal, silently drop raw ingress at stream capacity, or report
storage/runtime evidence without proving the actual route.

## Solution

- Bound replay preview, cascade archive handoffs, output validation, and
scale-test expectations while preserving complete processing.
- Preserve and validate material coordinate units before archive/replay.
- Prove manifest-backed CAS replay after source mutation/removal through
the real file-source production path.
- Make CAS fsck/work control explicit, resumable, cancellable, and
unlimited by default unless operators select limits.
- Retain manifest roots through ordinary cleanup and require explicit
purge authority.
- Backpressure the source-material JetStream stream at capacity instead
of silently discarding old raw frames.
- Make isolated daemon tests use durable workspace storage and include
actionable startup storage diagnostics.

## Verification

- `SINEX_TEST_TMPDIR=/realm/project/sinex/.sinex/test-tmp xtask test -p
sinexd -E 'test(/manifest_and_source_removal_obligation/)'
--impact-mode=off` — 1 passed (invocation 2007862).
- Bounded 10,001-root replay route — passed with complete cardinality,
bounded scan batches, and bounded archive handoff (recorded on
`sinex-p61n.1`).
- Focused CAS lifecycle, manifest-retention, fsck, and replay proofs —
passed; residual crash/power-loss and live-topology evidence remains
tracked in Beads.
- Pre-push schema-bundle check — passed.
- `xtask check --changed-strict origin/master` — 8 packages passed.

## Remaining campaign work

This PR does not close the wipe gate. Remaining dependencies include
intentional master-baseline re-verification, CAS crash/power-loss and
live deployment evidence, source/parser fidelity breadth, backup/restore
drills, work-control coverage, and the rest of the authoritative Beads
census. Phase C remains destructive and requires the milestone's
explicit operator authorization.

Ref: `sinex-p61n.1`, `sinex-r6d.14`, `sinex-r6d.16`, `sinex-kn14`.

---------

+++ crate/sinex-primitives/src/rpc/ops.rs
@@ -19,7 +19,7 @@ pub struct Operation {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub preview_summary: Option<Value>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub duration_ms: Option<i32>,
+    pub duration_ms: Option<i64>,
 }
 
 // ─────────────────────────────────────────────────────────────
